pyproject.toml
[project]
name = "autopoe"
authors = [{ name = "ImFeH2", email = "i@feh2.im" }]
dynamic = ["version"]
license = "Apache-2.0"
description = "A structured multi-agent framework for coordinated AI collaboration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.15.0",
    "fastapi[standard]",
    "httpx>=0.28.0",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[project.scripts]
autopoe = "app.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "app/_version.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "frontend/node_modules",
    "worktrees",
    "logs",
    "build",
    "dist",
    "*.spec",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
artifacts = ["app/static/**"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pyinstaller>=6.0",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "N", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
